backend/requirements.txt
Flask==3.0.2
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.6.0
pymysql==1.1.0
python-dotenv==1.0.0
Pillow==10.3.0
cryptography>=38.0
gunicorn==22.0.0
